Fixed type for y

+++ src/train/__init__.py
@@ -47,7 +47,7 @@ def train(
 
         # Split the data into training and validation sets
         x_train, x_val = x[train_index].float(), x[val_index].float()
-        y_train, y_val = y[train_index].float(), y[val_index].float()
+        y_train, y_val = y[train_index].type(torch.int), y[val_index].type(torch.int)
 
         # Define the train and validation dataloaders
         train_dataset = torch.utils.data.TensorDataset(x_train, y_train)
